commit
Added a new folder for the archaeological sites in the state of Morelos, called 'morelos-zones', also a new folder for the related images called 'morelos-img'.

Added the 8 archaeological sites located in the state of Morelos: Chalcatzingo, Coatetelco, Las Pilas, Olintepec, Teopanzolco, Tepozteco, Xochicalco and Yautepec, all with information and location.